[#27435] compilers.tvm: Added basic torch support & VM compilation
add basic torchconversion function to handle (tested on sample size n=1)
tracing and pushing the model through relay.frontend

add class variable that determines whether to compile for the relay.VM
or not.
For compatibility, the modelpath still takes a file as input, however
due to the VM compilation producing 2 files, it appends .so and .ro
(relay object) as extensions. This way, the only thing required to
change the compilation from vm-based to 'normal' is removal of a future
argument flag

Signed-off-by: Jan Bylicki <jbylicki@internships.antmicro.com>